Works on OpenCL 1.1 devices, V6_CPU

+++ v2.5 2020-10-12/g2_expr.cpp
@@ -131,7 +131,7 @@ void				prof_print()
 			GUIPrint(xpos2, ypos, "%lf", p.second);
 		//	GUIPrint(xpos2, ypos, "%g", p.second);
 			int xstart=xpos-10, ymiddle=ypos+8;
-			line(xstart, ymiddle, xstart-0.2*p.second, ymiddle);
+			line(xstart, ymiddle, int(xstart-0.2*p.second), ymiddle);
 		}
 		//GUIPrint(xpos, k*18, longest.first.c_str());
 		//GUIPrint(xpos2, k*18, "%lf", longest.second);
